requirements.txt
Flask>=3.0,<4.0
python-docx>=1.1,<2.0
lxml>=5.0,<6.0
PyMuPDF>=1.24,<2.0
